pyproject.toml
[project]
name = "solai-signals"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "orjson==3.10.7",
  "asyncpg==0.29.0",
  "SQLAlchemy==2.0.34",
  "alembic==1.13.2",
  "redis==5.0.8",
  "websockets==10.4",
  "aiohttp==3.10.3",
  "tenacity==9.0.0",
  "python-telegram-bot==21.6",
  "cryptography==43.0.1",
  "solana==0.34.3",
  "anchorpy==0.20.1",
  "numpy==2.1.1",
  "pandas==2.2.2",
  "httpx==0.27.2",
  "prometheus-client==0.20.0",
  "structlog==24.4.0",
]

[tool.setuptools]
packages = [
    "api", "bots", "bus", "connectors", "ops", "risk", "signals",
    "smartmoney", "storage", "trading", "workers"
]
